cleaning up ternaries into variables
